feat: implement internationalization for doctor profile components

+++ front-end/src/utils/date-time/formatDate.js
@@ -1,7 +1,7 @@
 import i18n from 'i18next';
 
 const formatDate = (date, config) => {
-    const defaultOptions = { day: 'numeric', month: 'long', year: 'numeric' };
+    const defaultOptions = { day: 'numeric', month: 'short', year: 'numeric' };
     const options = config || defaultOptions;
     const locale = i18n.language || 'en';
